backend/requirements.txt
altgraph==0.17.4
asgiref==3.11.0
attrs==25.4.0
autobahn==25.12.2
Automat==25.4.16
blinker==1.9.0
cbor2==5.8.0
certifi==2026.2.25
cffi==2.0.0
channels==4.3.2
chardet==5.2.0
chargily_pay==2.1.3
charset-normalizer==3.4.5
click==8.3.0
colorama==0.4.6
constantly==23.10.4
contourpy==1.3.3
cryptography==46.0.3
customtkinter==5.2.2
cycler==0.12.1
daphne==4.2.1
darkdetect==0.8.0
dj-database-url==3.1.2
Django==5.2.1
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
dotenv==0.9.9
Flask==3.1.2
Flask-SQLAlchemy==3.1.1
fonttools==4.60.1
greenlet==3.2.4
gunicorn==25.1.0
hyperlink==21.0.0
idna==3.11
Incremental==24.11.0
itsdangerous==2.2.0
Jinja2==3.1.6
kiwisolver==1.4.9
MarkupSafe==3.0.3
matplotlib==3.9.2
msgpack==1.1.2
numpy==2.3.4
packaging==25.0
pandas==2.2.3
pefile==2023.2.7
pillow==11.2.1
psutil==7.0.0
psycopg2-binary==2.9.11
py-ubjson==0.16.1
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pyinstaller==6.13.0
pyinstaller-hooks-contrib==2025.3
PyJWT==2.11.0
pyOpenSSL==25.3.0
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
pywin32-ctypes==0.2.3
qrcode==8.2
reportlab==4.2.5
requests==2.31.0
service-identity==24.2.0
setuptools==79.0.1
six==1.17.0
SQLAlchemy==2.0.43
sqlparse==0.5.3
tk==0.1.0
Twisted==25.5.0
txaio==25.12.2
typing_extensions==4.15.0
tzdata==2025.2
ujson==5.11.0
urllib3==2.6.3
Werkzeug==3.1.3
zope.interface==8.2
